% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{add_title_to_desc}
\alias{add_title_to_desc}
\title{Adding a title to a description}
\usage{
add_title_to_desc(desc, title_col = "repo")
}
\arguments{
\item{desc}{The return value of \code{\link[=read_description_files]{read_description_files()}}, i.e.
a tabular description of repos.}

\item{title_col}{The column name of \code{desc} that contains the title to use.}
}
\description{
Adding a title to a description
}
