% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{check_status}
\alias{check_status}
\title{Check a directory for uncomitted changes}
\usage{
check_status(dir = ".")
}
\arguments{
\item{dir}{A directory to check for unpushed changes.}
}
\description{
Check a directory for uncomitted changes
}
\seealso{
\code{\link[=team_check_unpushed]{team_check_unpushed()}}
}
