% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{check_unpushed_files}
\alias{check_unpushed_files}
\title{Check a directory for unpushed changes}
\usage{
check_unpushed_files(dir = ".")
}
\arguments{
\item{dir}{A directory to check for unpushed changes.}
}
\description{
Check a directory for unpushed changes
}
\seealso{
\code{\link[=team_check_uncommitted]{team_check_uncommitted()}}
}
