% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{communicate_remote}
\alias{communicate_remote}
\title{Communicate with a remote}
\usage{
communicate_remote(dir, fun = pull, credentials = NULL,
  communicate_result = TRUE)
}
\arguments{
\item{dir}{A directory.}

\item{fun}{A function to apply. Either pull or push.}

\item{credentials}{Credentials passed to \code{fun}, e.g. see \code{\link[git2r:pull]{git2r::pull()}}.}
}
\description{
Communicate with a remote
}
