% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_ext}
\alias{drop_ext}
\title{Drop everything after the last extension character}
\usage{
drop_ext(char, ext_sep = ".")
}
\arguments{
\item{char}{A character string.}

\item{ext_chr}{A character}
}
\description{
Drop everything after the last extension character
}
\examples{
teamtools:::drop_ext("a.b.csv")
}
