% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{find_repo_root}
\alias{find_repo_root}
\title{Find the root of a git repo}
\usage{
find_repo_root(...)
}
\arguments{
\item{...}{Arguments passed on to \code{find_root_from_dir}
\describe{
  \item{}{}
}}
}
\description{
Simly go up the tree until a \code{.git} folder is found.
}
