% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{find_root_from_dir}
\alias{find_root_from_dir}
\title{Find a root given a directory that must exist there}
\usage{
find_root_from_dir(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=find_root_from_object]{find_root_from_object}}
  \describe{
    \item{\code{root_object}}{A file or directory in a directory that qualifies the
latter as a root.}
    \item{\code{dir}}{The directory to start the search from.}
    \item{\code{object_finder}}{A function that is applied to the \code{root_object}. If
the root object is a file, use \code{\link[rprojroot:has_file]{rprojroot::has_file()}}, if it is a
directory, use \code{\link[rprojroot:has_dir]{rprojroot::has_dir()}}.}
  }}
}
\description{
Find a root given a directory that must exist there
}
