% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{find_root_from_object}
\alias{find_root_from_object}
\title{Find a root from an object}
\usage{
find_root_from_object(root_object, dir = ".", object_finder)
}
\arguments{
\item{root_object}{A file or directory in a directory that qualifies the
latter as a root.}

\item{dir}{The directory to start the search from.}

\item{object_finder}{A function that is applied to the \code{root_object}. If
the root object is a file, use \code{\link[rprojroot:has_file]{rprojroot::has_file()}}, if it is a
directory, use \code{\link[rprojroot:has_dir]{rprojroot::has_dir()}}.}
}
\description{
Finds a root from either a file or a directory.
}
