% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{find_team_root}
\alias{find_team_root}
\title{Find a team root}
\usage{
find_team_root(...)
}
\arguments{
\item{...}{Arguments passed on to \code{find_root_from_file}
\describe{
  \item{}{}
}}
}
\description{
Find a team root
}
