% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_dirs}
\alias{git_dirs}
\title{Get all git directories under a team root}
\usage{
git_dirs(dir = ".")
}
\arguments{
\item{dir}{Any directory under a team root.}
}
\description{
Get all git directories under a team root
}
