% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{ls_projects}
\alias{ls_projects}
\alias{ls_repos}
\title{List all projects / repos under a team root}
\usage{
ls_projects(dir = ".")

ls_repos(dir = ".")
}
\arguments{
\item{dir}{Any directory under the team root.}
}
\description{
List all projects / repos under a team root
}
