% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_repo}
\alias{open_repo}
\title{Open an RStudio Project from a repository}
\usage{
open_repo(repo = "meta", project = NULL, dir = ".", newSession = TRUE)
}
\arguments{
\item{repo}{A character vector of length one with a repository name.}

\item{project}{A character vector of length one with a project name.
Can be left \code{NULL} if the repo name is unique accros all projects.}

\item{dir}{Any directory under the team root.}

\item{newSession}{whether or not to open the RStudio Project in a new
session.}
}
\description{
Open an RStudio Project from a repository
}
