% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{read_and_markdown_description_files}
\alias{read_and_markdown_description_files}
\title{Read descirption files and convert them to markdown}
\usage{
read_and_markdown_description_files(projects = NULL)
}
\arguments{
\item{projects}{A character vector with projects to include. If \code{NULL},
all projects will be returned.}
}
\description{
Reads description files and put them from a tabular format into a
string that uses markdown syntax.
}
