% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_description_file}
\alias{read_description_file}
\title{Read a description file from a repo}
\usage{
read_description_file(project, repo)
}
\arguments{
\item{project}{The name of the project. Corresponds to a directory.}

\item{repo}{The name of the repo. Correcponds to a sub-directory of \code{project}.}
}
\description{
See \code{\link[=read_repo_file]{read_repo_file()}} for details. Arguments \code{file} and \code{root_file} already
prefilled.
}
