% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_description_files}
\alias{read_description_files}
\title{Read all description files into a tabular format}
\usage{
read_description_files(projects = NULL)
}
\arguments{
\item{projects}{A character vector with projects to include. If \code{NULL},
all projects will be returned.}
}
\description{
Read all description files into a tabular format
}
