% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_local_description}
\alias{read_local_description}
\title{Read a description file from a repo}
\usage{
read_local_description(simplify = TRUE)
}
\arguments{
\item{simplify}{Whether to collapse the lines into a long string.}
}
\description{
One element in the return value corresponds to one line in the description.
}
