% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_repo_file}
\alias{read_repo_file}
\title{Read a file from a repo given a root file}
\usage{
read_repo_file(project, repo, file, root_file = ".team")
}
\arguments{
\item{project}{The name of the project. Corresponds to a directory.}

\item{repo}{The name of the repo. Correcponds to a sub-directory of \code{project}.}

\item{file}{The file to read in.}

\item{root_file}{A file that will will be used as a root criterion. See
\code{\link[rprojroot:has_file]{rprojroot::has_file()}} for details.}
}
\description{
Read a file from a repo given a root file
}
