% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{read_team_config}
\alias{read_team_config}
\title{Read a team's configuration}
\usage{
read_team_config(dir = ".")
}
\arguments{
\item{dir}{Any directory under the root.}
}
\description{
Leverages the \code{yaml} package to read the configuration of the team.
}
