% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{repo_to_desc}
\alias{repo_to_desc}
\title{Adding the repo as a title to the description}
\usage{
repo_to_desc(desc)
}
\arguments{
\item{desc}{The return value of \code{\link[=read_description_files]{read_description_files()}}, i.e.
a tabular description of repos.}
}
\description{
Adding the repo as a title to the description
}
