% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_null_to}
\alias{set_null_to}
\title{Set a null value to a default}
\usage{
set_null_to(test, default)
}
\arguments{
\item{test}{That vale that should be asserted.}

\item{default}{The value that should be returned if \code{test} is \code{NULL}.}
}
\description{
Set a null value to a default
}
\keyword{internal}
