% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{structure}
\alias{structure}
\title{How to organize team work}
\description{
The hierarchical structure to use is:
}
\details{
\itemize{
\item team root (corresponds to a team name on BitBucket).
\item project root (corresponds to a project on BitBucket).
\item repo (corresponds to a repository on BitBucket).
}

Further:
\itemize{
\item Put a .team file in the team root. Content does not matter.
\item Put a .description file in every repo that you want to describe and add
a markdown-like description to it.
\item Paste the contents of description into the README using
\verb{r paste(readLines(".description"), collapse = '\\n')}.
}
}
