% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{team_check_status}
\alias{team_check_status}
\title{Check all repositories in a team for their git status}
\usage{
team_check_status(dir = ".")
}
\arguments{
\item{dir}{Any directory under the team root.}
}
\description{
Check all repositories in a team for their git status
}
\seealso{
\code{\link[=check_status]{check_status()}}
}
