% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{team_check_uncomitted}
\alias{team_check_uncomitted}
\title{Check all repositories in a team for uncommitted changes}
\usage{
team_check_uncomitted(dir = ".")
}
\arguments{
\item{dir}{Any directory under the team root.}
}
\description{
Check all repositories in a team for uncommitted changes
}
\seealso{
\code{\link[=check_unpushed_files]{check_unpushed_files()}}
}
