% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{team_pull}
\alias{team_pull}
\alias{team_push}
\title{Push to or pull from all remote repos under a team}
\usage{
team_pull(credentials = team_credentials(), dir = ".")

team_push(credentials = team_credentials(), dir = ".")
}
\arguments{
\item{credentials}{Credentials passed to \code{fun}, e.g. see \code{\link[git2r:pull]{git2r::pull()}}.}

\item{dir}{Any directory under a team root.}
}
\description{
Push to or pull from all remote repos under a team
}
