% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_team_repo}
\alias{init_team_repo}
\title{Init a teamtools repo}
\usage{
init_team_repo(dir = ".")
}
\arguments{
\item{dir}{The directory of the directory that is a repository.}
}
\description{
Does some initial house keeping, i.e:
}
\details{
\itemize{
\item adds a .description
\item adds a .Rprofile that will source the .Rprofile from the team root.
}
}
