% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_repo}
\alias{open_repo}
\title{Open an RStudio Project from a repository}
\usage{
open_repo(repo = "meta", n = 1, project = NULL, dir = ".", newSession = TRUE)
}
\arguments{
\item{repo}{A character vector of length one with a repository name. Fixed
regex (i.e. substring matches) are allowed.}

\item{n}{For partial matches, the ID of repos to open in parallel. Can be a
single integer, a sequence (\code{3:7}) or a vector (\code{c(2, 5)}). For \code{n = 1},
the exact match is opened.}

\item{project}{A character vector of length one with a project name.
Can be left \code{NULL} if the repo name is unique across all projects.}

\item{dir}{Any directory under the team root.}

\item{newSession}{whether or not to open the RStudio Project in a new
session.}
}
\description{
Open an RStudio Project from a repository
}
